# Hiccup parsing perf improvements