--- a/shadow-cljs.edn
+++ b/shadow-cljs.edn
@@ -31,8 +31,8 @@
   :benchmark {:target :node-script
               :main hx.benchmark/main
               :output-to "target/benchmark/index.js"
-              :devtools {:enabled false
-                         ;; :after-load hx.benchmark/main
+              :devtools {;; :enabled false
+                         :after-load hx.benchmark/main
                          }}
   :benchmark-browser {:target :browser
                       :output-dir "public/benchmark/js"