--- a/src/hx/react.cljs
+++ b/src/hx/react.cljs
@@ -139,4 +139,3 @@
   a new React element"
   [component]
   (partial $ component))
-