Before:
```
react x 1,468,464 ops/sec ±1.52% (57 runs sampled)
reagent x 325,920 ops/sec ±0.38% (62 runs sampled)
hx x 44,782 ops/sec ±0.65% (58 runs sampled)
```

Current After:
```
react x 1,720,359 ops/sec ±0.45% (62 runs sampled)
reagent x 337,924 ops/sec ±0.43% (61 runs sampled)
hx x 178,537 ops/sec ±0.51% (62 runs sampled)
```

hx is still lagging behind reagent by about 30%. Current goal is to flesh out benchmark with more cases, and be able to at least match reagent in most cases.

My dream would be to achieve ~50% of raw React, but that might require some tradeoffs that aren't worth it.