@Lokeh 2019-04-24
Latest changes are pretty good:

```
react x 1,528,503 ops/sec ±2.25% (58 runs sampled)
reagent x 318,871 ops/sec ±0.52% (62 runs sampled)
hx x 319,236 ops/sec ±1.76% (60 runs sampled)
```

Some improvements were made to converting props obj back to map as well, though this benchmark doesn't reflect that.

Validating different cases to make sure I'm not introducing any regressions.