@niwinz 2019-07-29
Have you considered 100% static compilation of hiccup (something like this: https://github.com/rauhs/hicada) I'm using it in my own "rum" [fork](https://github.com/funcool/rumext) (still WIP the final api, because is used on a big project and I need to make changes incrementally...) and it has almost 0% overhead on top of react.