@Lokeh 2019-07-29
@niwinz I have; I'm working on an experiment using reader tags https://github.com/Lokeh/thump

So far I enjoy it but it does have its tradeoffs. Will probably blog about it as I get closer to releasing to general use.